Move memory state operations to MemoryStateOperations.[hc]pp
