+++ jlm/llvm/ir/operators/operators.cpp
@@ -1178,54 +1178,6 @@ ExtractValue::copy() const
   return std::unique_ptr<rvsdg::operation>(new ExtractValue(*this));
 }
 
-/* MemStateMerge operator */
-
-MemStateMergeOperator::~MemStateMergeOperator()
-{}
-
-bool
-MemStateMergeOperator::operator==(const rvsdg::operation & other) const noexcept
-{
-  auto operation = dynamic_cast<const MemStateMergeOperator *>(&other);
-  return operation && operation->narguments() == narguments();
-}
-
-std::string
-MemStateMergeOperator::debug_string() const
-{
-  return "MemStateMerge";
-}
-
-std::unique_ptr<rvsdg::operation>
-MemStateMergeOperator::copy() const
-{
-  return std::unique_ptr<rvsdg::operation>(new MemStateMergeOperator(*this));
-}
-
-/* MemStateSplit operator */
-
-MemStateSplitOperator::~MemStateSplitOperator()
-{}
-
-bool
-MemStateSplitOperator::operator==(const rvsdg::operation & other) const noexcept
-{
-  auto operation = dynamic_cast<const MemStateSplitOperator *>(&other);
-  return operation && operation->nresults() == nresults();
-}
-
-std::string
-MemStateSplitOperator::debug_string() const
-{
-  return "MemStateSplit";
-}
-
-std::unique_ptr<rvsdg::operation>
-MemStateSplitOperator::copy() const
-{
-  return std::unique_ptr<rvsdg::operation>(new MemStateSplitOperator(*this));
-}
-
 /* malloc operator */
 
 malloc_op::~malloc_op()
